refactored pax-manifest
