+++ pax-compiler/Cargo.toml
@@ -33,6 +33,7 @@ tar = "0.4"
 fs_extra = "1.3.0"
 pax-message = {version = "0.10.7", path="../pax-message"}
 pax-runtime-api = {version= "0.10.7", path="../pax-runtime-api"}
+pax-manifest = {path="../pax-manifest"}
 pest = "2.6.0"
 pest_derive = "2.6.0"
 itertools = "0.10.5"
