+++ pax-compiler/src/expressions.rs
@@ -1,6 +1,7 @@
-use super::manifest::{
-    ComponentDefinition, ControlFlowRepeatPredicateDefinition, ExpressionSpec,
-    ExpressionSpecInvocation, PaxManifest, PropertyDefinition, TemplateNodeDefinition,
+use pax_manifest::{
+    escape_identifier, ComponentDefinition, ControlFlowRepeatPredicateDefinition, ExpressionSpec,
+    ExpressionSpecInvocation, PaxManifest, PropertyDefinition, PropertyDefinitionFlags,
+    SettingElement, SettingsBlockElement, TemplateNodeDefinition, Token, TypeDefinition, TypeTable,
     ValueDefinition,
 };
 use std::collections::HashMap;
@@ -9,10 +10,6 @@ use std::slice::IterMut;
 
 use crate::errors::source_map::SourceMap;
 use crate::errors::PaxTemplateError;
-use crate::manifest::{
-    PropertyDefinitionFlags, SettingElement, SettingsBlockElement, Token, TypeDefinition, TypeTable,
-};
-use crate::parsing::escape_identifier;
 use color_eyre::eyre;
 use color_eyre::eyre::Report;
 use lazy_static::lazy_static;
